+++ src/App/components/Topbar/index.js
@@ -1,25 +1,15 @@
 import React, { Fragment } from "react";
 import PropTypes from "prop-types";
-import classnames from "classnames";
+import logo from "$/img/logo/spay_horizontal_pos.svg";
 
-const TopbarBtn = ({ icon, text }) => {
-    const iconClass = classnames("material-icons", icon === "menu" ? "topbar-btn-icon" : "");
+const isDev = process.env.version === "LOCAL_DEV";
 
-    if (icon || text) {
-        return (
-            <button type="button" className="topbar-btn">{"\n\t\t"}
-                {icon ?
-                    <>
-                        <i className={iconClass}>{icon}</i>
-                        {"\n\t\t"}
-                    </>
-                    : null}
-            </button>
-        );
-    }
-
-    return null;
-};
+const TopbarBtn = () => (
+    <button type="button" className="topbar-btn">{"\n\t\t"}
+        <i className="material-icons topbar-btn-icon">menu</i>
+        {"\n\t\t"}
+    </button>
+);
 
 const TopbarMenu = ({ menu, logout }) => {
     const { items } = menu;
@@ -40,43 +30,40 @@ const TopbarMenu = ({ menu, logout }) => {
                         </a>
                     </Fragment>
                 ))}{"\n"}
-                {logout ?
-                    <>
-                        <a className="topbar-link-right" href="#" onClick={e => e.preventDefault()}>
-                            {"\n"}
-                            <i className="material-icons">exit_to_app</i>
-                            {"\n"}
-                            <span>Log out</span>
-                            {"\n"}
-                        </a>
-                        {"\n"}
-                    </> : null}
+                {logout ? <TopbarLogout /> : null}
             </div>
         </nav>
     );
 };
 
+const TopbarLogout = () => (
+    <>
+        <a className="topbar-link-right" href="#" onClick={e => e.preventDefault()}>{"\n"}
+            <i className="material-icons">exit_to_app</i>{"\n"}
+            <span>Log out</span>{"\n"}
+        </a>{"\n"}
+    </>
+);
+
+const TopbarLogo = () => (
+    <>
+        <a className="topbar-logo" href="/" onClick={e => e.preventDefault()}>{"\n"}
+            <img src={!isDev ? `${process.env.basename}img/spay_horizontal_pos.svg` : logo} alt="swedbankpay-logo"/>{"\n"}
+        </a>{"\n"}
+    </>
+);
+
 const Topbar = ({ topbarContent, wide, logout, id }) => (
     <header className={`topbar${wide ? ` topbar-${wide}-wide` : ""}`} id={id}>{"\n"}
         {topbarContent ?
         <>
-            <TopbarBtn icon={topbarContent.btn.icon} text={topbarContent.btn.name} target={topbarContent.id} />{"\n"}
-            {"\n"}<a href="#" className="topbar-logo" onClick={e => e.preventDefault()}></a>{"\n"}
+            <TopbarBtn />{"\n"}
+            <TopbarLogo />
             <TopbarMenu menu={topbarContent} logout={!!logout} />
         </> :
         <>
-            {"\n"} <a href="#" className="topbar-logo" onClick={e => e.preventDefault()}></a>{"\n"}
-            {logout ?
-                    <>
-                        <a className="topbar-link-right" href="#" onClick={e => e.preventDefault()}>
-                            {"\n"}
-                            <i className="material-icons">exit_to_app</i>
-                            {"\n"}
-                            <span>Log out</span>
-                            {"\n"}
-                        </a>
-                        {"\n"}
-                    </> : null}
+            <TopbarLogo />
+            {logout ? <TopbarLogout /> : null}
         </>}
     </header>
 );