+++ src/px-script/main/topbar/NavMenu.js
@@ -14,8 +14,6 @@ export default class NavMenu {
         this.linkContainer = this.navMenuElement.querySelector(".topbar-link-container");
         this.closeNavIcon = this.navMenuElement.querySelector(".close-topbar-nav");
         this.btnElement = topbarComponent.querySelector(SELECTORS.BTN);
-        this.iconElement = this.btnElement ? this.btnElement.querySelector(SELECTORS.ICON) : null;
-        this.userIcon = this.iconElement ? this.iconElement.innerHTML : null;
 
         if (this.btnElement) {
             this.btnElement.addEventListener("click", e => {
@@ -44,29 +42,29 @@ export default class NavMenu {
     _initAnchors () {
         // Closing menu for clicking on links in SPA's.
         this.navMenuElement.querySelectorAll("a")
-            .forEach(anchor => anchor.addEventListener("click", () => this.close()));
+            .forEach(anchor => anchor.addEventListener("click", () => {
+                if (this.isOpen) {
+                    this.close();
+                }
+            }));
     }
 
     open () {
         handleScrollbar();
         this.isOpen = true;
 
-        if (this.iconElement) { this.iconElement.innerHTML = "close"; }
-
+        this.navMenuElement.classList.add("topbar-nav-open");
         this.navMenuElement.classList.add("d-block");
-        setTimeout(() => {
-            this.navMenuElement.classList.add("topbar-nav-open");
-        }, 10);
     }
 
     close () {
         handleScrollbar();
         this.isOpen = false;
 
-        if (this.iconElement) { this.iconElement.innerHTML = this.userIcon; }
-
         this.navMenuElement.classList.remove("topbar-nav-open");
+        this.navMenuElement.classList.add("topbar-nav-closing");
         setTimeout(() => {
+            this.navMenuElement.classList.remove("topbar-nav-closing");
             this.navMenuElement.classList.remove("d-block");
         }, 300);
     }