+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 ### Changed
 
 - Changed typography sizes and added breakpoint styling for headings/small.
+- Changed topbar logo to be an image wrapped by an anchor to be more accessible.
 
 ### Fixed
+
+- Fixed issue with esc event listener for topbar nav menu.
+- Updated animation for mobile topbar to make it smoother.
 
 ### Removed